backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coreschema==0.0.4
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
setuptools==75.6.0
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
